requirements.txt
python-telegram-bot==21.6
groq==0.11.0
anthropic==0.34.2
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.10.5
